add macro crate dependency and example that uses it - add ragout.rs module and move lib.rs contents there - add lib and custom_events exclusive features - add cfg for code chunks that would use each of the two features
